requirements.txt
django
numpy
pyyaml
tqdm
pathos
opencv-python-headless
pillow
